Add capability for fetch-fixes to write complete MachineConfigs
Before, we would write the raw MachineConfigs on disk, which don't
contain names nor metadata as those are filled by the Compliace
Operator.

With these change these could be used directly instead of asking
users to edit them. This is done by adding the needed labels and name to
the objects, based on the contents of a new flag introduced called
`--mc-roles`.

Signed-off-by: Juan Antonio Osorio Robles <jaosorior@redhat.com>