fetch-fixes: Add support for adding extra manifests
This adds the concept of extra manifest builders for the `fetch-fixes`
subcommand. The main use-case is that some GitOps systems will require
extra manifests to function appropriately.

e.g.

MachineConfigs don't play well with ArgoCD, so we need an extra job in
the form of hooks to pause/unpause the MachineConfigPool.

this takes the form of the `--manifest-prepare` flag, which may be
`default` (doing nothing) or `ArgoCD` (adding the manifests necessary
for ArgoCD).

Note that this PR does very basic dependency resolution... but only goes
1 level deep. We need to make this more dynamic in the future.

Signed-off-by: Juan Antonio Osorio Robles <jaosorior@redhat.com>