Document the branches and their purpose
We recently updated the `stable` branch and realized we didn't have any
process written down for how the `stable` branch is updated. Updating
`stable` is important for addresssing issues in underlying go
dependencies.

This commit documents the branches and briefly describes that the
`stable` branch should be rebased on top of `master` to pick up new
changes.