Use cp.Complete()
Recently changes to cp.Run() require us to pass the src and dest
arguments for cp into the command with `Complete()` instead of with
`Run()`.

This commit updates our code to be compatible with kubeclt 0.26.1. It
also adds an important note about a workaround we need for cp to work
with the Kubernetes Client struct we use.

Since we're using `Complete()` to build the command, we must override
the group version, API path, and negotiation variables after we
`Complete()` the command. If we set those variables before calling
`Complete()`, they'll be overwritten by the implementation of
`Complete()` and result in an error when client-go attempts to make the
API call.

See https://github.com/operator-framework/operator-sdk/issues/1570 for
more information.