Update k8s dependencies to 0.28.2
Update all the k8s dependencies together. Sometimes when we, or
dependabot, attempt to update dependencies separately, we get build
errors because k8s libraries are expected to run with the same semantic
version. This is noted in https://github.com/kubernetes/client-go/issues/1269

Let's group the k8s dependencies into a single patch so we reduce the
possibility of build issues due to compatibility issues between client
libraries.